Add world gen package
