+++ core/src/com/travall/game/WorldLoading.java
@@ -1,0 +1,7 @@
+package com.travall.game;
+
+import com.badlogic.gdx.ScreenAdapter;
+
+public class WorldLoading extends ScreenAdapter {
+	
+}
